REFACTOR: Separate `Completions` from `InlineInput` for reuse.
