+++ packages/frontend/src/components/index.ts
@@ -1,3 +1,4 @@
+export * from "./completions";
 export * from "./icon_button";
 export * from "./inline_input";
 export * from "./rich_text_editor";
