+++ packages/frontend/src/components/inline_input.css
@@ -39,53 +39,12 @@
     border-bottom: 3px solid indianred;
 }
 
-.completion-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.completion-list li {
-    cursor: pointer;
-    padding: 5px;
-}
-
-.completion-list li.active {
-    background: #f1f5f9;
-}
-
-.completion-head {
-    display: flex;
-    gap: 2em;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.completion-description {
-    font-size: 0.75em;
-    color: dimgray;
-}
-
-.completion-shortcut {
-    font-size: 0.9em;
-}
-
-.completion-empty {
-    font-size: 0.75em;
-    color: dimgray;
-}
-
-kbd.shortcut {
-    display: flex;
-    gap: 0.5em
-}
-
-.completions-container {
+.inline-input-completions {
     position: absolute;
     z-index: 10;
     background: white;
     padding: 0.25em;
     border: solid dimgray 1px;
     border-radius: 4px;
-    box-shadow: 3px 3px 3px gray;
+    box-shadow: 3px 3px 3px rgba(0, 0, 0, 0.5);
 }
